+++ lib/rhales/utils.rb
@@ -1,5 +1,12 @@
 # lib/rhales/utils.rb
 
+module Rhales
+  module Utils
+    # Utility modules and classes
+  end
+end
+
 require_relative 'utils/json_serializer'
 require_relative 'utils/schema_generator'
 require_relative 'utils/schema_extractor'
+require_relative 'utils/logging_helpers'